third_party/blink/web_tests/fast/dom/shadow/selection-shouldnt-expose-shadow-dom-expected.txt

PASS internals.treeScopeRootNode(selection.anchorNode) is document
PASS internals.treeScopeRootNode(selection.focusNode) is document
PASS successfullyParsed is true

TEST COMPLETE
